Fixed many warnings (compare (un)signed, unused vars, and initialize with zero)
git-svn-id: svn://svn.lighttpd.net/lighttpd/branches/lighttpd-1.4.x@2160 152afb58-edef-0310-8abb-c4023f1b3aa9

+++ src/plugin.c
@@ -197,7 +197,7 @@ int plugins_load(server *srv) {
 
 #else
 #if 1
-		init = (int (*)(plugin *))dlsym(p->lib, srv->tmp_buf->ptr);
+		init = (int (*)(plugin *))(intptr_t)dlsym(p->lib, srv->tmp_buf->ptr);
 #else
 		*(void **)(&init) = dlsym(p->lib, srv->tmp_buf->ptr);
 #endif
